Create a wrapper type for user messages
