+++ test/re_web/graphql/messages_test.exs
@@ -118,11 +118,13 @@ defmodule ReWeb.GraphQL.MessagesTest do
       query = """
         {
           listingUserMessages (listingId: #{listing.id}) {
-            id
-            inserted_at
-            message
-            listing {
+            messages {
               id
+              inserted_at
+              message
+              listing {
+                id
+              }
             }
           }
         }
@@ -137,11 +139,11 @@ defmodule ReWeb.GraphQL.MessagesTest do
         post(conn, "/graphql_api", AbsintheHelpers.query_skeleton(query, "listingUserMessages"))
 
       assert %{
-               "listingUserMessages" => [
+               "listingUserMessages" => %{"messages" => [
                  %{"id" => ^id1, "listing" => %{"id" => ^listing_id}, "inserted_at" => _},
                  %{"id" => ^id2, "listing" => %{"id" => ^listing_id}, "inserted_at" => _},
                  %{"id" => ^id3, "listing" => %{"id" => ^listing_id}, "inserted_at" => _}
-               ]
+               ]}
              } = json_response(conn, 200)["data"]
     end
 
@@ -163,10 +165,12 @@ defmodule ReWeb.GraphQL.MessagesTest do
       query = """
         {
           listingUserMessages (listingId: #{listing.id}) {
-            id
-            message
-            listing {
+            messages {
               id
+              message
+              listing {
+                id
+              }
             }
           }
         }
@@ -181,11 +185,11 @@ defmodule ReWeb.GraphQL.MessagesTest do
         post(conn, "/graphql_api", AbsintheHelpers.query_skeleton(query, "listingUserMessages"))
 
       assert %{
-               "listingUserMessages" => [
+               "listingUserMessages" => %{"messages" => [
                  %{"id" => ^id1, "listing" => %{"id" => ^listing_id}},
                  %{"id" => ^id2, "listing" => %{"id" => ^listing_id}},
                  %{"id" => ^id3, "listing" => %{"id" => ^listing_id}}
-               ]
+               ]}
              } = json_response(conn, 200)["data"]
     end
 
@@ -206,13 +210,15 @@ defmodule ReWeb.GraphQL.MessagesTest do
       query = """
         {
           listingUserMessages (senderId: #{user1.id}) {
-            id
-            message
-            sender {
+            messages {
               id
-            }
-            listing {
-              id
+              message
+              sender {
+                id
+              }
+              listing {
+                id
+              }
             }
           }
         }
@@ -227,14 +233,14 @@ defmodule ReWeb.GraphQL.MessagesTest do
         post(conn, "/graphql_api", AbsintheHelpers.query_skeleton(query, "listingUserMessages"))
 
       assert %{
-               "listingUserMessages" => [
+               "listingUserMessages" => %{"messages" => [
                  %{
                    "id" => ^id1,
                    "listing" => %{"id" => ^listing_id},
                    "sender" => %{"id" => ^user1_id}
                  },
                  %{"id" => ^id2, "sender" => %{"id" => ^user1_id}}
-               ]
+               ]}
              } = json_response(conn, 200)["data"]
     end
 
@@ -258,13 +264,15 @@ defmodule ReWeb.GraphQL.MessagesTest do
       query = """
         {
           listingUserMessages (listingId: #{listing.id}, senderId: #{user1.id}) {
-            id
-            message
-            sender {
+            messages {
               id
-            }
-            listing {
-              id
+              message
+              sender {
+                id
+              }
+              listing {
+                id
+              }
             }
           }
         }
@@ -278,13 +286,13 @@ defmodule ReWeb.GraphQL.MessagesTest do
         post(conn, "/graphql_api", AbsintheHelpers.query_skeleton(query, "listingUserMessages"))
 
       assert %{
-               "listingUserMessages" => [
+               "listingUserMessages" => %{"messages" => [
                  %{
                    "id" => ^id1,
                    "listing" => %{"id" => ^listing_id},
                    "sender" => %{"id" => ^user1_id}
                  }
-               ]
+               ]}
              } = json_response(conn, 200)["data"]
     end
 
@@ -292,8 +300,9 @@ defmodule ReWeb.GraphQL.MessagesTest do
       query = """
         {
           listingUserMessages {
-            id
-            message
+            messages {
+              id
+            }
           }
         }
       """
